refactor(routes): update imports to pages/meter and remove stale water/ stubs
- Update all route files to import Meter and MeterMap from pages/meter/
  instead of the old pages/water/ location
- Delete pages/water/Meter.tsx and pages/water/MeterMap.tsx (empty stubs
  superseded by the new pages/meter/ implementations)
