--- frontend/src/pages/water/Meter.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const MeterLocationManagement: React.FC = () => {
-    return (
-        <div>
-            Hello Meter Manage
-        </div>
-    );
-};
-
-export default MeterLocationManagement;
