--- frontend/src/pages/water/MeterMap.tsx
@@ -1,11 +1,0 @@
-
-const MeterMap = () => {
-    return (
-      <div>
-        Hello Meter Map
-    </div>
-  );
-};
-
-
-export default MeterMap;
