+++ frontend/src/routes/UserRoutes.tsx
@@ -2,7 +2,7 @@ import { RouteObject } from "react-router-dom";
 import SidebarLayout from "@/layout/SidebarLayout";
 import Water from "../pages/water/Water";
 import Profile from "../pages/profile/ProfilePage";
-import WaterMeterMap from "../pages/water/MeterMap";
+import WaterMeterMap from "../pages/meter/MeterMap";
 import NotiMaintainLog from "../pages/maintain/NotiMaintainPage";
 import NotificationPage from "../pages/notification/NotificationPage";
 import NotFound from "../pages/404/404Page";
